Исправил замечания, произвел декомпозицию задачи
