+++ react-chat/src/components/chat-send-message-form/index.css
@@ -1,0 +1,45 @@
+.form {
+    width: 100%;
+    display: inline-block;
+    border: 2px solid rgba(25, 25, 25, 0.32);
+    transition: background-color 0.1s ease, box-shadow 0.1s ease, transform 0.1s ease;
+
+    display: flex;
+    padding: 5px;
+    background-color: white;
+    border-top: 1px solid #DAAD86;
+    height: 50px;
+}
+
+.form:active {
+    background-color: rgba(50, 50, 50, 0.7);
+    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.2);
+    transform: scale(1.01);
+}
+
+.form-input {
+    flex: 1;
+    border: 1px solid #DAAD86;
+    border-radius: 5px;
+    padding: 10px;
+    outline: none;
+}
+
+.attach-button, .send-button {
+    background: none;
+    border: none;
+    cursor: pointer;
+    padding-left: 10px;
+    transform: scale(2);
+}
+
+.attach-button:hover, .send-button:hover {
+    filter: brightness(0.1);
+    opacity: 2;
+}
+
+input {
+    border: 10px;
+    outline: none;
+    width: calc(100% - 2px);
+}
