+++ react-chat/src/components/header/HeaderChat.css
@@ -1,0 +1,105 @@
+.chat-header {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    background-color: #659DBD;
+    color: white;
+    padding: 10px 15px;
+    height: 100px;
+}
+
+.back-button {
+    transform: scale(2);
+    margin-right: 10px;
+}
+
+.back-button:hover {
+    filter: brightness(0.7);
+    opacity: 0.8;
+}
+
+.profile-link {
+    margin-right: 50px;
+    margin-left: 50px;
+    transform: scale(3);
+    transition: transform 0.5s ease;
+}
+
+.profile-link:hover {
+    transform: scale(2.1);
+}
+
+.user-info {
+    margin-right: 70px;
+}
+
+.user-info h3 {
+    display: flex;
+    flex-direction: column;
+    margin: 0 0 10px 0;
+    text-align: center;
+    cursor: pointer;
+    transition: color 0.5s ease-in;
+}
+
+.user-info p {
+    margin: 0;
+    text-align: center;
+    font-size: 12px;
+    color: #ddd;
+}
+
+.user-info h3:hover {
+    color: #aaa;
+}
+
+.right-icons {
+    display: flex;
+    gap: 10px;
+    margin-left: auto;
+}
+
+.right-icons button{
+    background: none;
+    border: none;
+    cursor: pointer;
+    transform: scale(2);
+}
+
+.right-icons button:hover {
+    filter: brightness(0.7);
+    opacity: 0.8;
+}
+
+@media (max-width: 768px) {
+    .chat-header {
+        height: 60px;
+    }
+
+    .back-button {
+        transform: scale(1.5);
+    }
+
+    .profile-link {
+        transform: scale(2);
+        margin-right: 15px;
+        margin-left: 5px;
+    }
+
+    .user-info h3 {
+        font-size: 15px;
+        margin-bottom: 0px;
+    }
+
+    .user-info p {
+        font-size: 10px;
+    }
+
+    .right-icons button {
+        transform: scale(1.5);
+    }
+
+    .user-info {
+        margin-right: 0px;
+    }
+}
