+++ react-chat/src/components/header/HeaderChatList.css
@@ -1,0 +1,66 @@
+.messenger-header {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    background-color: #659DBD;
+    color: white;
+    padding: 10px 15px;
+    height: 100px;
+}
+
+.burger {
+    background: none;
+    border: none;
+    cursor: pointer;
+    transform: scale(2);
+    margin-right: 10px;
+}
+
+.burger:hover {
+    filter: brightness(0.7);
+}
+
+.messenger {
+    margin-right: 50px;
+    margin-left: 50px;
+}
+
+.search-icon {
+    margin-left: auto;
+    background: none;
+    border: none;
+    cursor: pointer;
+    transform: scale(2);
+}
+
+.search-icon:hover {
+    filter: brightness(0.7);
+    opacity: 0.8;
+}
+
+@media (max-width: 768px) {
+    .messenger-header {
+        height: 60px;
+    }
+
+    .burger {
+        background: none;
+        border: none;
+        cursor: pointer;
+        transform: scale(1.5);
+        margin-right: 10px;
+    }
+
+    .messenger {
+        margin-right: 50px;
+        margin-left: 20px;
+    }
+
+    .search-icon {
+        transform: scale(1.5);
+    }
+
+    .messenger h1 {
+        font-size: 20px;
+    }
+}
