+++ react-chat/src/components/header/index.jsx
@@ -1,0 +1,2 @@
+export {HeaderChat} from './HeaderChat.jsx';
+export {HeaderChatList} from './HeaderChatList.jsx';
