--- react-chat/src/pages/Chat.css
@@ -1,224 +1,0 @@
-html {
-    box-sizing: border-box;
-    overflow: hidden;
-}
-
-*,
-*::before,
-*::after {
-    box-sizing: inherit;
-}
-
-body {
-    margin: 0;
-    background-color: #fff;
-    color: #333;
-    font-size: 15px;
-    font-family: Helvetica, Arial, sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    overflow: hidden;
-}
-
-.form {
-    width: 100%;
-    display: inline-block;
-    border: 2px solid rgba(25, 25, 25, 0.32);
-    transition: background-color 0.1s ease, box-shadow 0.1s ease, transform 0.1s ease;
-}
-
-.form:active {
-    background-color: rgba(50, 50, 50, 0.7);
-    box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.2);
-    transform: scale(1.01);
-}
-
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    background-color: #659DBD;
-    color: white;
-    padding: 10px 15px;
-    height: 100px;
-}
-
-.back-button {
-    transform: scale(2);
-    margin-right: 10px;
-}
-
-.back-button:hover {
-    filter: brightness(0.7);
-    opacity: 0.8;
-}
-
-.profile-link {
-    margin-right: 50px;
-    margin-left: 50px;
-    transform: scale(3);
-    transition: transform 0.5s ease;
-}
-
-.profile-link:hover {
-    transform: scale(2.1);
-}
-
-.user-info {
-    margin-right: 70px;
-}
-
-.user-info h3 {
-    display: flex;
-    flex-direction: column;
-    margin: 0 0 10px 0;
-    text-align: center;
-    cursor: pointer;
-    transition: color 0.5s ease-in;
-}
-
-.user-info p {
-    margin: 0;
-    text-align: center;
-    font-size: 12px;
-    color: #ddd;
-}
-
-.user-info h3:hover {
-    color: #aaa;
-}
-
-
-.right-icons {
-    display: flex;
-    gap: 10px;
-    margin-left: auto;
-}
-
-.right-icons button{
-    background: none;
-    border: none;
-    cursor: pointer;
-    transform: scale(2);
-}
-
-.right-icons button:hover {
-    filter: brightness(0.7);
-    opacity: 0.8;
-}
-
-.chat-container {
-    display: flex;
-    flex-grow: 1;
-    flex-direction: column;
-    height: calc(100vh - 150px);
-    overflow-y: auto;
-    overflow-x: hidden;
-    padding: 10px;
-    background-color: #FBEEC1;
-}
-
-@keyframes slide-in-left {
-    0% {
-        opacity: 0; 
-        transform: translateX(-20px); 
-    }
-    100% {
-        opacity: 1; 
-        transform: translateX(0); 
-    }
-}
-
-@keyframes slide-in-right {
-    0% {
-        opacity: 0; 
-        transform: translateX(20px); 
-    }
-    100% {
-        opacity: 1; 
-        transform: translateX(0); 
-    }
-}
-
-.message.opponent {
-    color: #8D8741;
-    border: 1px solid;
-    border-radius: 4px;
-    margin-bottom: 10px;
-    padding: 4px;
-
-    background-color: #DAAD86;
-    border-radius: 10px;
-    padding: 10px;
-    max-width: 70%;
-    align-self: flex-start;
-    word-wrap: break-word;
-
-    animation: slide-in-left 0.5s forwards;
-}
-
-.message.you {
-    color: #8D8741;
-    border: 1px solid;
-    border-radius: 4px;
-    margin-bottom: 10px;
-    padding: 4px;
-
-    background-color: #DAAD86;
-    border-radius: 10px;
-    padding: 10px;
-    max-width: 70%;
-    align-self: flex-end;
-    word-wrap: break-word;
-
-    animation: slide-in-right 0.5s forwards;
-}
-
-.form {
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    display: flex;
-    align-items: center;
-    padding: 5px;
-    background-color: white;
-    border-top: 1px solid #DAAD86;
-    height: 50px;
-
-}
-
-.form-input {
-    flex: 1;
-    border: 1px solid #DAAD86;
-    border-radius: 5px;
-    padding: 10px;
-    outline: none;
-}
-
-input[type=submit] {
-    visibility: collapse;
-}
-
-.message-text {
-    color: black;
-}
-
-.attach-button, .send-button {
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding-left: 10px;
-    transform: scale(2);
-}
-
-.attach-button:hover, .send-button:hover {
-    filter: brightness(0.1);
-    opacity: 2;
-}
-
-input {
-    border: 10px;
-    outline: none;
-    width: calc(100% - 2px);
-}
