+++ react-chat/src/pages/chat/index.css
@@ -1,0 +1,38 @@
+html {
+    box-sizing: border-box;
+    overflow: hidden;
+}
+
+*,
+*::before,
+*::after {
+    box-sizing: inherit;
+}
+
+body {
+    margin: 0;
+    background-color: #fff;
+    color: #333;
+    font-size: 15px;
+    font-family: Helvetica, Arial, sans-serif;
+    font-weight: 400;
+    line-height: 20px;
+    overflow: hidden;
+}
+
+.chat-container {
+    display: flex;
+    flex-grow: 1;
+    flex-direction: column;
+    height: calc(100vh - 150px);
+    overflow-y: auto;
+    overflow-x: hidden;
+    padding: 10px;
+    background-color: #FBEEC1;
+}
+
+@media (max-width: 768px) {
+    .chat-container {
+        height: calc(100vh - 110px);
+    }
+}
