+++ react-chat/src/pages/messenger/index.css
@@ -18,47 +18,6 @@ body {
     line-height: 20px;
 }
 
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    background-color: #659DBD;
-    color: white;
-    padding: 10px 15px;
-    height: 100px;
-}
-
-.burger {
-    background: none;
-    border: none;
-    cursor: pointer;
-    transform: scale(2);
-    margin-right: 10px;
-}
-
-.burger:hover {
-    filter: brightness(0.7);
-    opacity: 0.8;
-}
-
-.messanger {
-    margin-right: 50px;
-    margin-left: 50px;
-}
-
-.search-icon {
-    margin-left: auto;
-    background: none;
-    border: none;
-    cursor: pointer;
-    transform: scale(2);
-}
-
-.search-icon:hover {
-    filter: brightness(0.7);
-    opacity: 0.8;
-}
-
 .window {
     display: flex;
     flex-direction: column;
@@ -72,6 +31,8 @@ body {
     flex-grow: 1;
     overflow-y: auto;
     background-color: #FBEEC1;
+
+    word-wrap: break-word;
 }
 
 @keyframes pulse {
@@ -91,17 +52,28 @@ body {
 
 .start-chat-button {
     position: fixed;
-    z-index: 100;
     bottom: 30px;
     right: 30px;
+    width: 25px;
+    height: 25px;
     border: none;
     background-color: #659DBD;
-    border-radius: 100%;
-    transform: scale(2.5);
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50%;
+
+    /* transform: scale(2.5); */
     font-size: 25px;
     cursor: pointer;
 
     animation: pulse 5s infinite;
+}
+
+.start-chat-button {
+    max-width: 80%;
+    max-height: 80%;
 }
 
 .start-chat-button:hover {
@@ -113,9 +85,17 @@ body {
     display: flex;
 }
 
+.button-container {
+    display: flex;
+    justify-content: flex-end;
+}
+
 .close-user-selection-button {
     border: none;
     background: none;
+    position: absolute;
+    top: 5px;
+    right: 0;
 }
 
 .close-user-selection-button:hover {
@@ -123,7 +103,18 @@ body {
     opacity: 0.8;
 }
 
+.overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5);
+}
+
 .users-container {
+    border-radius: 10px;
+    padding: 20px;
     position: fixed;
     top: 50%;
     left: 50%;
@@ -137,27 +128,37 @@ body {
     flex-direction: column;
 }
 
-#chatList{
-    list-style-type: none;
+#chatList {
+    padding: 0;
     width: 300px;
-    max-height: 400px;
+    max-height: 350px;
     overflow-y: auto;
+    list-style-type: none;
 }
 
 #chatList li {
     display: flex;
-    align-items: center;
     padding: 10px;
+    cursor: pointer;
+    border-bottom: 1px solid #8E592C;
+}
+
+#chatList li:hover {
+    background-color: #A38A74;
 }
 
 #chatList a {
     color: inherit;
+    justify-content: flex-start;
     display: flex;
+    align-items: center;
+    margin: 10px;
+
+    text-decoration: none;
 }
 
 .user-icon {
     transform: scale(1.9);
-    margin-bottom: 50px;
     margin-right: 20px;
 }
